// ngApp/src/app/Models/special-item.ts
export class SpecialItem {
    Id: number = 0;
    CodeId = 0;
    CodeName: string = "";
    Acronym: string = "";
    StructureId: number = 0;
    Name: string = "";
    STime: number = 0;
    ETime: number = 0;
    StartTime: string = "";
    EndTime: string = "";
    Title: string = "";
    Type: string = "";
    enable: boolean = false;
    color:string = "";
    icon:string = "";
    
    PersonId: number = 0;    
    Date: string ="";
}